solana-agent-kit-py-main/pyproject.toml
[tool.poetry]
name = "solana-agent-kit-py"
version = "1.4.0"
description = "connect any ai agents to solana protocols"
authors = ["sendaifun <dev@sendaifun.com>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/sendaifun/solana-agent-kit-py"
repository = "https://github.com/sendaifun/solana-agent-kit-py"
keywords = ["solana", "sdk", "blockchain", "tools"]
packages = [
    { include = "solana_agent_kit", from = "." },
]


[tool.poetry.dependencies]
python = "^3.13.0"
requests = "2.32.3"
base58 = "^2.1.1"
pillow = "^11.0.0"
openai = "^1.58.1"
pydantic = "^2.10.4"
pre-commit = "^4.0.1"
sphinx = "^8.1.3"
sphinx-rtd-theme = "^3.0.2"
sphinx-autodoc-typehints = "^2.5.0"
setuptools = "^75.6.0"
langchain = "^0.3.12"
wheel = "^0.45.1"
twine = "^6.0.1"
numpy = "^1.26.2,<3"
python-dotenv = "^1.0.1"
aiohttp = "^3.11.11"
solders = "^0.21.0,<0.24.0"
solana = "^0.35.0"
anchorpy = "^0.20.1"
pythclient = "^0.2.0"
pydantic-ai = ">=0.0.19"
cryptography = "^44.0.0"
pynacl = "^1.5.0"
backpack-exchange-sdk = "^1.0.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dev-dependencies]
pytest = "8.3.4"
black = "24.10.0"
isort = "^5.10.0"
